Lesson on symbols
Followed along the lesson on symbols. However, the video was severely
lacking -- it did not present any motivation as to why symbols would in
any way be useful. Luckily, there is a great article by Jason Orendorff
of Mozilla explaining a possible use case for symbols. The article can
be found at https://hacks.mozilla.org/2015/06/es6-in-depth-symbols/